Implemented insert, retrieve and report gRpc endpoints

+++ Intra-Block.Cache/Report.cs
@@ -2,9 +2,9 @@ namespace Intra_Block.Cache
 {
     public class Report
     {
-        public int NumberOfEntries { get; set; }
-        public int CurrentMemoryUsage { get; set; }
-        public int Uptime { get; set; }
+        public ulong NumberOfEntries { get; set; }
+        public ulong CurrentMemoryUsage { get; set; }
+        public ulong Uptime { get; set; }
         public ConnectedClient Clients { get; set; }
         public Averages Averages { get; set; }
